hanif => perbaiki response + fix bugs

+++ app/Http/Controllers/Api/SellerController.php
@@ -3,24 +3,22 @@
 namespace App\Http\Controllers\Api;
 
 use App\Cart;
-use App\User;
+use App\Shop;
 use App\Product;
 use GuzzleHttp\Client;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
 use App\Http\Controllers\Controller;
-use App\Shop;
 
 class SellerController extends Controller
 {
     public function store(Request $request, Client $client)
     {
-        // Cek jika dia bukan pemilik toko
+        // Cek jika dia tidak memiliki toko
         $cek = Shop::find(Auth::id());
 
         if (empty($cek)) {
-            return $this->SendResponse('failed', "You don't have permission to create product", null, 404);
+            return $this->SendResponse('failed', "You don't have permission to create product", null, 400);
         }
 
         // validate
@@ -64,27 +62,14 @@ class SellerController extends Controller
 
     public function update(Request $request, Client $client, $id)
     {
-        // cek jika dia bukan pemilik toko
-        $cek = Shop::where('id', '!=', Auth::id());
+        // cek jika dia bukan pemilik toko dan produk yang dicari tidak ada
+        $product = Product::where('id', $id)->where('shop_id', Auth::id())->first();
 
-        $product = Product::find($id);
-
-        if ($product) {
-            return $product;
+        if (empty($product)) {
+            return $this->SendResponse('failed', "You don't have permission to update this product", null, 400);
         }
 
-        echo "kosong";
-
-        die;
-
-        if ($product == false) {
-            return $this->SendResponse('failed', 'Product not found', null, 404);
-        }
-
-        return $product;
-
-        die;
-
+        // validasi
         $request->validate([
             'product_name' => 'required|min:10|max:60',
             'price'        => 'required',
@@ -94,31 +79,48 @@ class SellerController extends Controller
         ]);
 
         // Validasi image
-        $image = base64_encode(file_get_contents($request->image));
+        if ($request->image) {
+            $image = base64_encode(file_get_contents($request->image));
 
-        $res = $client->request('POST', 'https://freeimage.host/api/1/upload', [
-            'form_params' => [
-                'key' => '6d207e02198a847aa98d0a2a901485a5',
-                'action' => 'upload',
-                'source' => $image,
-                'format' => 'json'
-            ]
+            $res = $client->request('POST', 'https://freeimage.host/api/1/upload', [
+                'form_params' => [
+                    'key' => '6d207e02198a847aa98d0a2a901485a5',
+                    'action' => 'upload',
+                    'source' => $image,
+                    'format' => 'json'
+                ]
+            ]);
+
+            $get = $res->getBody()->getContents();
+
+            $hasil = json_decode($get);
+
+            $newImage = $hasil->image->display_url;
+        } else {
+            $newImage = $product->image;
+        }
+
+        $data = $product->update([
+            'product_name' => $request->product_name,
+            'price' => $request->price,
+            'quantity' => $request->quantity,
+            'description' => $request->description,
+            'image' => $newImage,
         ]);
 
-        $get = $res->getBody()->getContents();
-
-        $hasil = json_decode($get);
-
-        $newAvatar = $hasil->image->display_url;
-
-        DB::update();
-
-        // return $this->SendResponse('success', 'Produk berhasil diubah', $data, 201);
+        return $this->SendResponse('success', 'Product updated successfully', $data, 200);
     }
 
     public function destroy($id)
     {
-        // hapus keranjang dengan id produk
+        // jika dia bukan pemilik toko dan id yang dicari tidak ada
+        $product = Product::where('id', $id)->where('shop_id', Auth::id())->first();
+
+        if (empty($product)) {
+            return $this->SendResponse('failed', "You don't have permission to delete this product", null, 400);
+        }
+
+        // hapus keranjang dengan yang memiliki produk berkait
         Cart::where('product_id', $id)->delete();
 
         $data = Product::destroy($id);
@@ -126,8 +128,10 @@ class SellerController extends Controller
         if ($data) {
             return response([
                 'status' => 'success',
-                'message' => 'Data berhasil dihapus'
+                'message' => 'Your product deleted successfully'
             ], 200);
-        } else return $this->SendResponse('failed', 'Produk gagal dihapus', null, 404);
+        } else {
+            return $this->SendResponse('failed', 'Error salahkan backend', null, 500);
+        }
     }
 }
